Updating to BigNumber for most numbers

+++ core/src/main/kotlin/com/dllewellyn/coinbaseapi/models/currency/OpenOrder.kt
@@ -3,19 +3,20 @@ package com.dllewellyn.coinbaseapi.models.currency
 import com.dllewellyn.coinbaseapi.models.BuyOrSell
 import com.dllewellyn.coinbaseapi.models.trade.LimitOrMarket
 import com.dllewellyn.coinbaseapi.models.trade.OrderStatus
+import java.math.BigInteger
 
 data class OpenOrder(
     val createdAt: String,
-    val executedValue: Double,
-    val fillFees: Double,
-    val fillSize: Double,
+    val executedValue: BigInteger,
+    val fillFees: BigInteger,
+    val fillSize: BigInteger,
     val id: String,
     val postOnly: Boolean,
-    val price: Double,
+    val price: BigInteger,
     val currencyPair: CurrencyPair,
     val settled: Boolean,
     val side: BuyOrSell,
-    val size: Double,
+    val size: BigInteger,
     val status: OrderStatus,
     val stp: String?,
     val timeInForce: String,
